Moved lblDungeon to separate QWidget
