+++ GameController/QToolboxActions.py
@@ -1,5 +1,5 @@
 from PyQt5 import QtWidgets
-from PyQt5.QtWidgets import QHBoxLayout, QBoxLayout, QVBoxLayout, QPushButton, QFormLayout, QLabel,QWidget
+from PyQt5.QtWidgets import QHBoxLayout, QBoxLayout, QVBoxLayout, QPushButton, QFormLayout, QLabel, QWidget
 from PyQt5.QtCore import Qt
 
 
@@ -9,12 +9,12 @@ class QToolboxActions(QtWidgets.QWidget):
         self.layoutMainVer = QFormLayout()
         self.infoLbl = QLabel()
         self.actions = []
-        self.buttons_size = parent.toolbar_w
+        self.requested_size, _ = parent.get_toolbar_size()
         self.initUI()
         # self.setAttribute(QtCore.Qt.WA_StyledBackground, True)
 
     def initUI(self):
-        w,h=self.buttons_size-10,self.buttons_size-10
+        w, h = self.requested_size - 10, self.requested_size - 10
         button_wait = QPushButton("Wait")
         button_wait.setStyleSheet("background-color: rgb(233, 185, 110);")
         button_wait.setFixedSize(w, h)
@@ -31,7 +31,7 @@ class QToolboxActions(QtWidgets.QWidget):
         self.layoutMainVer.addWidget(button_wait)
         self.layoutMainVer.addWidget(button_walk)
         self.layoutMainVer.addWidget(button_tap)
-        #self.layoutMainVer.setVerticalSpacing(10)
-        #self.layoutMainVer.setHorizontalSpacing(0)
+        # self.layoutMainVer.setVerticalSpacing(10)
+        # self.layoutMainVer.setHorizontalSpacing(0)
         self.layoutMainVer.setContentsMargins(0, 5, 0, 0)
         self.setLayout(self.layoutMainVer)
